+++ deploy/operator/Makefile
@@ -4,7 +4,7 @@ IMG ?= controller:latest
 
 # ENVTEST_K8S_VERSION refers to the version of kubebuilder assets to be downloaded by envtest binary.
 ENVTEST_K8S_VERSION = 1.30.0
-ENVTEST_PACKAGES ?= ./internal/controller
+ENVTEST_PACKAGES ?= ./internal/controller ./internal/webhook/validation
 ENVTEST_ASSETS_CMD = $(ENVTEST) use $(ENVTEST_K8S_VERSION) --bin-dir $(LOCALBIN) -p path
 
 # Get the currently used golang install path (in GOPATH/bin, unless GOBIN is set)
