+++ deploy/operator/internal/controller/testing/grove.io/clustertopologybindings.yaml
@@ -1,0 +1,20 @@
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  name: clustertopologybindings.grove.io
+spec:
+  group: grove.io
+  names:
+    kind: ClusterTopologyBinding
+    listKind: ClusterTopologyBindingList
+    plural: clustertopologybindings
+    singular: clustertopologybinding
+  scope: Cluster
+  versions:
+  - name: v1alpha1
+    served: true
+    storage: true
+    schema:
+      openAPIV3Schema:
+        type: object
+        x-kubernetes-preserve-unknown-fields: true
