+++ deploy/operator/internal/webhook/validation/dynamographdeploymentrequest_test.go
@@ -1,300 +1,19 @@
 /*
- * SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
  */
 
 package validation
 
 import (
-	"slices"
 	"strings"
 	"testing"
 
-	nvidiacomv1alpha1 "github.com/ai-dynamo/dynamo/deploy/operator/api/v1alpha1"
 	nvidiacomv1beta1 "github.com/ai-dynamo/dynamo/deploy/operator/api/v1beta1"
-	"github.com/ai-dynamo/dynamo/deploy/operator/internal/consts"
-	"github.com/ai-dynamo/dynamo/deploy/operator/internal/features"
 	admissionv1 "k8s.io/api/admission/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
-
-func TestDynamoGraphDeploymentRequestValidator_Validate(t *testing.T) {
-	requestValidators := requestValidatorsFromCRD(t, "nvidia.com_dynamographdeploymentrequests.yaml")
-
-	tests := []struct {
-		name          string
-		request       runtime.Object
-		oldRequest    runtime.Object
-		gpuDiscovery  bool
-		wantSchemaErr string
-		wantCELErr    string
-		wantWebhook   []string
-		wantWarnings  []string
-	}{
-		// Source-version schema, CEL, and conversion boundaries.
-		{
-			name:         "valid v1beta1 request",
-			request:      betaDGDRForAdmission(nil),
-			gpuDiscovery: true,
-		},
-		{
-			name:         "valid v1alpha1 request converts through the production path",
-			request:      alphaDGDRForAdmission(nil),
-			gpuDiscovery: true,
-		},
-		{
-			name: "v1beta1 empty model is rejected by source schema",
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Model = ""
-			}),
-			gpuDiscovery:  true,
-			wantSchemaErr: `spec.model: Invalid value: "": spec.model in body should be at least 1 chars long`,
-		},
-		{
-			name: "v1beta1 SLA optimization enum is rejected by source schema",
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				optimization := nvidiacomv1beta1.OptimizationType("cost")
-				request.Spec.SLA = &nvidiacomv1beta1.SLASpec{OptimizationType: &optimization}
-			}),
-			gpuDiscovery:  true,
-			wantSchemaErr: `spec.sla.optimizationType: Unsupported value: "cost": supported values: "latency", "throughput"`,
-		},
-		{
-			name: "v1alpha1 backend enum is rejected by source schema before conversion",
-			request: alphaDGDRForAdmission(func(request *nvidiacomv1alpha1.DynamoGraphDeploymentRequest) {
-				request.Spec.Backend = "unknown"
-			}),
-			gpuDiscovery:  true,
-			wantSchemaErr: `spec.backend: Unsupported value: "unknown": supported values: "auto", "vllm", "sglang", "trtllm"`,
-		},
-
-		// Structural create rules.
-		{
-			name: "DGD-only metadata annotations are ignored",
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Annotations = map[string]string{consts.KubeAnnotationDynamoOperatorOriginVersion: "not-semver"}
-			}),
-			gpuDiscovery: true,
-		},
-		{
-			name: "thorough search requires a concrete backend",
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Backend = nvidiacomv1beta1.BackendTypeAuto
-				request.Spec.SearchStrategy = nvidiacomv1beta1.SearchStrategyThorough
-			}),
-			gpuDiscovery: true,
-			wantWebhook: []string{
-				`spec.searchStrategy: Invalid value: "thorough": is incompatible with spec.backend "auto"; set spec.backend to a specific backend (sglang, trtllm, or vllm)`,
-			},
-		},
-		{
-			name:         "GPU discovery permits omitted hardware",
-			request:      betaDGDRForAdmission(nil),
-			gpuDiscovery: true,
-		},
-		{
-			name:    "disabled GPU discovery requires manual hardware",
-			request: betaDGDRForAdmission(nil),
-			wantWebhook: []string{
-				"spec.hardware: Required value: GPU hardware configuration is required when GPU discovery is disabled",
-			},
-		},
-		{
-			name: "manual hardware permits disabled GPU discovery",
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Hardware = &nvidiacomv1beta1.HardwareSpec{GPUSKU: nvidiacomv1beta1.GPUSKUTypeH100SXM}
-			}),
-		},
-		{
-			name: "independent create failures aggregate in API declaration order",
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Backend = nvidiacomv1beta1.BackendTypeAuto
-				request.Spec.SearchStrategy = nvidiacomv1beta1.SearchStrategyThorough
-			}),
-			wantWebhook: []string{
-				"spec.hardware: Required value: GPU hardware configuration is required when GPU discovery is disabled",
-				`spec.searchStrategy: Invalid value: "thorough": is incompatible with spec.backend "auto"; set spec.backend to a specific backend (sglang, trtllm, or vllm)`,
-			},
-		},
-
-		// Structural update rules.
-		{
-			name: "unchanged spec is accepted during profiling",
-			oldRequest: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Status.Phase = nvidiacomv1beta1.DGDRPhaseProfiling
-			}),
-			request:      betaDGDRForAdmission(nil),
-			gpuDiscovery: true,
-		},
-		{
-			name: "spec update is rejected during profiling",
-			oldRequest: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Status.Phase = nvidiacomv1beta1.DGDRPhaseProfiling
-			}),
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Model = alternateAdmissionModel
-			}),
-			gpuDiscovery: true,
-			wantWebhook: []string{
-				`spec: Forbidden: updates are forbidden while the resource is in phase "Profiling"; delete and recreate the resource to change its spec`,
-			},
-		},
-		{
-			name: "spec update is rejected during deploying",
-			oldRequest: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Status.Phase = nvidiacomv1beta1.DGDRPhaseDeploying
-			}),
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Model = alternateAdmissionModel
-			}),
-			gpuDiscovery: true,
-			wantWebhook: []string{
-				`spec: Forbidden: updates are forbidden while the resource is in phase "Deploying"; delete and recreate the resource to change its spec`,
-			},
-		},
-		{
-			name: "spec update is rejected during deployed",
-			oldRequest: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Status.Phase = nvidiacomv1beta1.DGDRPhaseDeployed
-			}),
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Model = alternateAdmissionModel
-			}),
-			gpuDiscovery: true,
-			wantWebhook: []string{
-				`spec: Forbidden: updates are forbidden while the resource is in phase "Deployed"; delete and recreate the resource to change its spec`,
-			},
-		},
-		{
-			name: "spec update is accepted during failed phase",
-			oldRequest: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Status.Phase = nvidiacomv1beta1.DGDRPhaseFailed
-			}),
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Model = alternateAdmissionModel
-			}),
-			gpuDiscovery: true,
-		},
-		{
-			name: "unchanged thorough and auto violation is ratcheted on update",
-			oldRequest: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Backend = nvidiacomv1beta1.BackendTypeAuto
-				request.Spec.SearchStrategy = nvidiacomv1beta1.SearchStrategyThorough
-			}),
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Backend = nvidiacomv1beta1.BackendTypeAuto
-				request.Spec.SearchStrategy = nvidiacomv1beta1.SearchStrategyThorough
-				request.Labels = map[string]string{"updated": "true"}
-			}),
-			gpuDiscovery: true,
-		},
-		{
-			name:       "missing hardware is ratcheted when GPU discovery becomes disabled",
-			oldRequest: betaDGDRForAdmission(nil),
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Labels = map[string]string{"updated": "true"}
-			}),
-		},
-		{
-			name: "removing manual hardware is rejected while GPU discovery is disabled",
-			oldRequest: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Hardware = &nvidiacomv1beta1.HardwareSpec{GPUSKU: nvidiacomv1beta1.GPUSKUTypeH100SXM}
-			}),
-			request: betaDGDRForAdmission(nil),
-			wantWebhook: []string{
-				"spec.hardware: Required value: GPU hardware configuration is required when GPU discovery is disabled",
-			},
-		},
-		{
-			name:       "newly introduced search violation is rejected on update",
-			oldRequest: betaDGDRForAdmission(nil),
-			request: betaDGDRForAdmission(func(request *nvidiacomv1beta1.DynamoGraphDeploymentRequest) {
-				request.Spec.Backend = nvidiacomv1beta1.BackendTypeAuto
-				request.Spec.SearchStrategy = nvidiacomv1beta1.SearchStrategyThorough
-			}),
-			gpuDiscovery: true,
-			wantWebhook: []string{
-				`spec.searchStrategy: Invalid value: "thorough": is incompatible with spec.backend "auto"; set spec.backend to a specific backend (sglang, trtllm, or vllm)`,
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			current := admissionUnstructured(t, tt.request)
-			var old map[string]any
-			if tt.oldRequest != nil {
-				old = admissionUnstructured(t, tt.oldRequest)
-			}
-
-			version := admissionSourceVersion(t, tt.request)
-			if tt.oldRequest != nil && admissionSourceVersion(t, tt.oldRequest) != version {
-				t.Fatal("old and current source versions differ")
-			}
-			requestValidator, ok := requestValidators[version]
-			if !ok {
-				t.Fatalf("no request validator for source version %q", version)
-			}
-
-			schemaErrs := requestValidator.validateSchema(current, old)
-			if tt.wantSchemaErr != "" {
-				if tt.wantCELErr != "" || len(tt.wantWebhook) != 0 || len(tt.wantWarnings) != 0 {
-					t.Fatal("schema rejection cannot have downstream expectations")
-				}
-				assertRequestValidationError(t, schemaErrs, tt.wantSchemaErr)
-				return
-			}
-			if len(schemaErrs) != 0 {
-				t.Fatalf("schema errors = %v, want none", schemaErrs)
-			}
-
-			celErrs := requestValidator.celValidator(current, old)
-			if tt.wantCELErr != "" {
-				if len(tt.wantWebhook) != 0 || len(tt.wantWarnings) != 0 {
-					t.Fatal("CEL rejection cannot have webhook expectations")
-				}
-				assertRequestValidationError(t, celErrs, tt.wantCELErr)
-				return
-			}
-			if len(celErrs) != 0 {
-				t.Fatalf("CEL errors = %v, want none", celErrs)
-			}
-
-			handler := NewDynamoGraphDeploymentRequestHandler()
-			ctx := dgdAdmissionContext(dgdrAdmissionOperation(tt.oldRequest), nvidiacomv1beta1.DynamoGraphDeploymentRequestGVK)
-			ctx = features.WithGate(ctx, features.Gates{GPUDiscovery: tt.gpuDiscovery})
-			var warnings []string
-			var err error
-			if tt.oldRequest == nil {
-				warnings, err = handler.ValidateCreate(ctx, dgdrAdmissionBeta(t, tt.request))
-			} else {
-				warnings, err = handler.ValidateUpdate(
-					ctx,
-					dgdrAdmissionBeta(t, tt.oldRequest),
-					dgdrAdmissionBeta(t, tt.request),
-				)
-			}
-			assertWebhookErrors(t, err, tt.wantWebhook)
-			if !slices.Equal(warnings, tt.wantWarnings) {
-				t.Fatalf("webhook warnings = %v, want %v", warnings, tt.wantWarnings)
-			}
-		})
-	}
-}
 
 func TestDynamoGraphDeploymentRequestHandlerBoundaryErrorsRemainRegular(t *testing.T) {
 	handler := NewDynamoGraphDeploymentRequestHandler()
@@ -306,79 +25,4 @@ func TestDynamoGraphDeploymentRequestHandlerBoundaryErrorsRemainRegular(t *testi
 	if k8serrors.IsInvalid(err) {
 		t.Fatalf("ValidateCreate() error = %v, want regular boundary error", err)
 	}
-}
-
-func betaDGDRForAdmission(
-	mutate func(*nvidiacomv1beta1.DynamoGraphDeploymentRequest),
-) *nvidiacomv1beta1.DynamoGraphDeploymentRequest {
-	request := &nvidiacomv1beta1.DynamoGraphDeploymentRequest{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: nvidiacomv1beta1.GroupVersion.String(),
-			Kind:       "DynamoGraphDeploymentRequest",
-		},
-		ObjectMeta: metav1.ObjectMeta{Name: "test-dgdr", Namespace: "default"},
-		Spec: nvidiacomv1beta1.DynamoGraphDeploymentRequestSpec{
-			Model:          "Qwen/Qwen3-0.6B",
-			Backend:        nvidiacomv1beta1.BackendTypeVllm,
-			Image:          "profiler:latest",
-			SearchStrategy: nvidiacomv1beta1.SearchStrategyRapid,
-		},
-	}
-	if mutate != nil {
-		mutate(request)
-	}
-	return request
-}
-
-func alphaDGDRForAdmission(
-	mutate func(*nvidiacomv1alpha1.DynamoGraphDeploymentRequest),
-) *nvidiacomv1alpha1.DynamoGraphDeploymentRequest {
-	request := &nvidiacomv1alpha1.DynamoGraphDeploymentRequest{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: nvidiacomv1alpha1.GroupVersion.String(),
-			Kind:       "DynamoGraphDeploymentRequest",
-		},
-		ObjectMeta: metav1.ObjectMeta{Name: "test-dgdr", Namespace: "default"},
-		Spec: nvidiacomv1alpha1.DynamoGraphDeploymentRequestSpec{
-			Model:   "Qwen/Qwen3-0.6B",
-			Backend: "vllm",
-			ProfilingConfig: nvidiacomv1alpha1.ProfilingConfigSpec{
-				ProfilerImage: "profiler:latest",
-			},
-		},
-	}
-	if mutate != nil {
-		mutate(request)
-	}
-	return request
-}
-
-func dgdrAdmissionBeta(
-	t *testing.T,
-	request runtime.Object,
-) *nvidiacomv1beta1.DynamoGraphDeploymentRequest {
-	t.Helper()
-	if request == nil {
-		return nil
-	}
-	switch request := request.(type) {
-	case *nvidiacomv1beta1.DynamoGraphDeploymentRequest:
-		return request.DeepCopy()
-	case *nvidiacomv1alpha1.DynamoGraphDeploymentRequest:
-		beta := &nvidiacomv1beta1.DynamoGraphDeploymentRequest{}
-		if err := request.ConvertTo(beta); err != nil {
-			t.Fatalf("convert v1alpha1 DGDR to v1beta1: %v", err)
-		}
-		return beta
-	default:
-		t.Fatalf("unsupported DGDR type %T", request)
-		return nil
-	}
-}
-
-func dgdrAdmissionOperation(oldRequest runtime.Object) admissionv1.Operation {
-	if oldRequest == nil {
-		return admissionv1.Create
-	}
-	return admissionv1.Update
 }
